refactor: auth 로직 분리 및 파일 경로 이동
