+++ src/router/router.js
@@ -1,4 +1,4 @@
-import { isAuthenticated } from "./auth";
+import { isAuthenticated } from "../auth/auth";
 
 export default function Router() {
   const routes = {};
